--- storage/interface.go
@@ -1,7 +1,0 @@
-// storage/interface.go
-package storage
-
-type StorageInterface interface {
-	Save(URL string) string
-	Resolve(id string) (string, error)
-}
